virtual board + viewport
The “Ant” now uses a virtual 32x32 board, the BitMatrix.  Viewport
starts at (0,0), the Ant starts randomly.  Wait a while or search
around with the arrows to find the Ant.

The world is a torus, the viewport is not — viewport can “jump” all the
way across, so it can be a little odd when you find an edge.

Arduino is a little silly, some of the debugging statements (still in
place) might show where I got ripped up.  Note that bool[32][32] is
stored as bytes, and smashes the stack.  malloc() pretty easily
destroys heap, and itself comes with a lot of otherwise-unnecessary
code.